+++ TheMillionaireGame/GameModule/View/GameView.swift
@@ -8,11 +8,9 @@
 import SwiftUI
 
 struct GameView: View {
-    @StateObject private var viewModel: GameViewModel
-    
-    init(viewModel: GameViewModel = GameViewModel()) {
-        _viewModel = StateObject(wrappedValue: viewModel)
-    }
+    @EnvironmentObject var viewModel: QuizViewModel
+    @Binding var currentScreen: MainScreenDestination
+
     
     var body: some View {
         ZStack{
@@ -20,7 +18,7 @@ struct GameView: View {
                 .edgesIgnoringSafeArea(.all)
             
             VStack(alignment: .center, spacing: 40){
-                HeaderView()
+                HeaderView(currentScreen: $currentScreen)
                 TimerView()
                     .environmentObject(viewModel)
                 VStack(spacing: 60){
@@ -62,46 +60,10 @@ struct GameView: View {
                 await viewModel.loadQuestions()
             }
         }
-        .fullScreenCover(isPresented: $viewModel.showPriceList) {
-            PriceListView(currentQuestion: viewModel.numberCurrentQuestion)
-        }
-        .navigationDestination(isPresented: $viewModel.isGameOver) {
-            HomeView()
-        }
-        .navigationBarHidden(true)
     }
 }
 
 #Preview{
-    //GameView()
-    let mockViewModel = GameViewModel()
-    mockViewModel.questions = [
-        Question(
-            type: "multiple",
-            difficulty: "easy",
-            category: "Entertainment: Film",
-            question: "What is the name of the protagonist in 'Cast Away'?",
-            correctAnswer: "Wilson",
-            incorrectAnswers: ["Carson", "Jackson", "Willy"]
-        ),
-        Question(
-            type: "multiple",
-            difficulty: "medium",
-            category: "Entertainment: Film",
-            question: "Which movie was NOT released in 1996?",
-            correctAnswer: "Gladiator",
-            incorrectAnswers: ["Independence Day", "The Rock", "Mission: Impossible"]
-        ),
-        Question(
-            type: "multiple",
-            difficulty: "hard",
-            category: "Entertainment: Film",
-            question: "Which animated movie first featured a celebrity voice actor?",
-            correctAnswer: "Aladdin",
-            incorrectAnswers: ["Toy Story", "James and the Giant Peach", "The Hunchback of Notre Dame"]
-        )
-    ]
-    mockViewModel.isLoading = false
-    
-    return GameView(viewModel: mockViewModel)
+    GameView(currentScreen: .constant(.game))
+        .environmentObject(QuizViewModel())
 }
