+++ TheMillionaireGame/HomeModule/View/HomeView.swift
@@ -1,6 +1,10 @@
 import SwiftUI
 
 struct HomeView: View {
+    
+    @EnvironmentObject var viewModel: QuizViewModel
+    @Binding var currentScreen: MainScreenDestination
+    
     @State private var showGameView = false
     @State private var showContinueGameView = false
     // MARK: - UI Constants
@@ -107,7 +111,7 @@ struct HomeView: View {
                         width: UI.Button.width,
                         height: UI.Button.height,
                         action: {
-                            showGameView = true
+                            currentScreen = .game
                         },
                         isLeadingText: false
                     )
@@ -118,17 +122,13 @@ struct HomeView: View {
             }
             .toolbar {
                 ToolbarItem(placement: .navigationBarTrailing) {
-                    Button(action: {}) {
+                    Button(action: {
+                        currentScreen = .rules
+                    }) {
                         Image(UI.HelpIcon.name)
                             .renderingMode(.original)
                     }
                 }
-            }
-            .navigationDestination(isPresented: $showContinueGameView) {
-                GameView()
-            }
-            .navigationDestination(isPresented: $showGameView) {
-                GameView()
             }
         }
     }
@@ -136,6 +136,7 @@ struct HomeView: View {
 
 struct HomeView_Previews: PreviewProvider {
     static var previews: some View {
-        HomeView()
+        HomeView(currentScreen: .constant(.home))
+            .environmentObject(QuizViewModel())
     }
 }
