--- TheMillionaireGame/ViewModels/ViewModel.swift
@@ -1,8 +1,0 @@
-//
-//  ViewModel.swift
-//  TheMillionaireGame
-//
-//  Created by Валентин on 20.07.2025.
-//
-
-import Foundation
