--- TheMillionaireGame/Views/ContentView.swift
@@ -1,26 +1,0 @@
-//
-//  ContentView.swift
-//  TheMillionaireGame
-//
-//  Created by Валентин on 20.07.2025.
-//
-
-import SwiftUI
-
-struct ContentView: View {
-    var body: some View {
-        VStack {
-            Image(systemName: "globe")
-                .imageScale(.large)
-                .foregroundStyle(.tint)
-            Text("Hello, world!")
-        }
-        
-        
-        .padding()
-    }
-}
-
-#Preview {
-    ContentView()
-}
